Bug 1278214 - ensure operation result is floating point. r=mats
MozReview-Commit-ID: AwIH4G2GlPv

UltraBlame original commit: ffe52673664fa60184c9221003b67ffad3465ca7

+++ layout/base/nsCSSRendering.cpp
@@ -3334,7 +3334,7 @@ static nscoord
 ComputeSpacedRepeatSize(nscoord aImageDimension,
                         nscoord aAvailableSpace,
                         bool& aRepeat) {
-  float ratio = aAvailableSpace / aImageDimension;
+  float ratio = static_cast<float>(aAvailableSpace) / aImageDimension;
 
   if (ratio < 2.0f) { 
     aRepeat = false;
